backend/requirements.txt
fastapi
uvicorn[standard]
pandas
psycopg2-binary
python-dotenv
requests
kafka-python
scikit-learn
xgboost
pytz
joblib